refactor: introduce UnifiedConfigurationModule aggregator
Centralize unified configuration wiring so each Standalone app no longer
re-lists UnifiedConfiguration, UnifiedConfigurationHelper, and every
*Override individually. UnifiedConfigurationModule is a single
@Configuration with @ComponentScan("io.camunda.configuration") that
picks all of them up; each override self-gates via @Profile or
@ConditionalOnSecondaryStorageType, so only the relevant beans
materialize for the active profiles.

OperatePropertiesOverride and TasklistPropertiesOverride now declare
@Profile("operate") / @Profile("tasklist") to make their loading
explicit. Affected unit tests gained the matching @ActiveProfiles.

ActorClock/IdleStrategy overrides remain un-gated because their
consumers in io.camunda.application.commons load under all profiles.

Standalone* and TestStandalone* apps now reference
UnifiedConfigurationModule.class instead of long source lists.
