+++ operate/common/src/main/java/io/camunda/operate/property/OperateProperties.java
@@ -16,8 +16,7 @@ import org.springframework.boot.context.properties.NestedConfigurationProperty;
  * This class contains all project configuration parameters.
  *
  * <p>NOTE: Some of the fields of this object are overridden with values coming from the Unified
- * Configuration system, from the object
- * io.camunda.configuration.beanoverrides.OperatePropertiesOverride
+ * Configuration system, from the object io.camunda.operate.OperatePropertiesOverride
  */
 public class OperateProperties {
