+++ tasklist/common/src/main/java/io/camunda/tasklist/property/TasklistProperties.java
@@ -14,8 +14,7 @@ import org.springframework.boot.context.properties.NestedConfigurationProperty;
  * This class contains all project configuration parameters.
  *
  * <p>NOTE: Some of the fields of this object are overridden with values coming from the Unified
- * Configuration system, from the object
- * io.camunda.configuration.beanoverrides.TasklistPropertiesOverride
+ * Configuration system, from the object io.camunda.tasklist.TasklistPropertiesOverride
  */
 public class TasklistProperties {
