+++ tasklist/qa/integration-tests/src/test/java/io/camunda/tasklist/management/HealthCheckIT.java
@@ -16,8 +16,8 @@ import static org.mockito.Mockito.verifyNoInteractions;
 
 import io.camunda.configuration.UnifiedConfiguration;
 import io.camunda.configuration.UnifiedConfigurationHelper;
-import io.camunda.configuration.beanoverrides.TasklistPropertiesOverride;
 import io.camunda.tasklist.JacksonConfig;
+import io.camunda.tasklist.TasklistPropertiesOverride;
 import io.camunda.tasklist.connect.ElasticsearchConnector;
 import io.camunda.tasklist.util.apps.nobeans.TestApplicationWithNoBeans;
 import java.util.Map;
